fix rew
