--- packages/frontend/src/shared/ui/agreements-list/index.tsx
@@ -1,1 +1,0 @@
-export * from './ui/AgreementsList'
