+++ packages/frontend/src/widgets/documents-agreements/ui/Agreements/Agreements.module.scss
@@ -3,7 +3,7 @@
   flex-direction: column;
   margin: 25px 0 333px;
 
-  &_contract {
+  .contract {
     width: 844px;
     margin-bottom: 40px;
   }
